+++ lib/SerializedCache.php
@@ -2,10 +2,26 @@
 
 namespace Amp\Cache;
 
+use Amp\Failure;
 use Amp\Promise;
+use Amp\Serialization\SerializationException;
+use Amp\Serialization\Serializer;
+use function Amp\call;
 
-interface SerializedCache
+final class SerializedCache
 {
+    /** @var Cache */
+    private $cache;
+
+    /** @var Serializer */
+    private $serializer;
+
+    public function __construct(Cache $cache, Serializer $serializer)
+    {
+        $this->cache = $cache;
+        $this->serializer = $serializer;
+    }
+
     /**
      * The promise is resolved with the unserialized cached value.
      *
@@ -14,33 +30,49 @@ interface SerializedCache
      * @return Promise<mixed|null> Resolves to the cached value nor `null` if it doesn't exist or fails with a
      * CacheException on failure.
      */
-    public function get(string $key): Promise;
+    public function get(string $key): Promise
+    {
+        return call(function () use ($key) {
+            $data = yield $this->cache->get($key);
+            if ($data === null) {
+                return $data;
+            }
+
+            return $this->serializer->unserialize($data);
+        });
+    }
 
     /**
      * Allows any serializable value to cached, not just strings.
      *
-     * @param $key string Cache key.
+     * @param $key   string Cache key.
      * @param $value mixed Value to cache.
-     * @param $ttl int Timeout in seconds. The default `null` $ttl value indicates no timeout. Values less than 0 MUST
-     * throw an \Error.
+     * @param $ttl   int Timeout in seconds. The default `null` $ttl value indicates no timeout. Values less than 0 MUST
+     *               throw an \Error.
      *
      * @return Promise<void> Resolves either successfully or fails with a CacheException on failure.
      */
-    public function set(string $key, $value, int $ttl = null): Promise;
+    public function set(string $key, $value, int $ttl = null): Promise
+    {
+        try {
+            $value = $this->serializer->serialize($value);
+        } catch (SerializationException $exception) {
+            return new Failure($exception);
+        }
+
+        return $this->cache->set($key, $value, $ttl);
+    }
 
     /**
      * Deletes a value associated with the given key if it exists.
-     *
-     * Implementations SHOULD return boolean `true` or `false` to indicate whether the specified key existed at the time
-     * the delete operation was requested. If such information is not available, the implementation MUST resolve the
-     * promise with `null`.
-     *
-     * Implementations MUST transparently succeed operations for non-existent keys.
      *
      * @param $key string Cache key.
      *
      * @return Promise<bool> Resolves to `true` / `false` to indicate whether the key existed or fails with a
      * CacheException on failure.
      */
-    public function delete(string $key): Promise;
+    public function delete(string $key): Promise
+    {
+        return $this->cache->delete($key);
+    }
 }
